requirements.txt
Cartridge==0.9.5
Django==1.6.8
Mezzanine==3.1.10
MySQL-python==1.2.3
PIL==1.1.7
Pillow==2.6.1
PyPDF2==1.23
beautifulsoup4==4.1.3
bleach==1.4
django-audit-log==0.5.0
django-daterange-filter==1.0.0
django-location-field==1.5.2
django-mockups==0.4.8
docutils==0.7
egenix-mx-base==3.1.3
filebrowser-safe==0.3.6
future==0.9.0
grappelli-safe==0.3.13
html5lib==1.0b3
numpy==1.5.0
oauthlib==0.6.3
phonenumbers==7.0.0
psycopg2==2.4.4
pycrypto==2.3
pytz==2014.7
reportlab==2.4
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
sorl-thumbnail==11.12
tzlocal==1.0
wsgiref==0.1.2
xhtml2pdf==0.0.6
